+++ src/lib/Events.ts
@@ -1,4 +1,5 @@
 import type { Intersection } from 'three'
+import type { IntersectionEvent } from '@threlte/extras'
 
 export type ThreeEvent<TSourceEvent> = Intersection & {
   nativeEvent: TSourceEvent
@@ -13,20 +14,22 @@ export type EventMap = {
   pointer: ThreeEvent<PointerEvent>
 }
 
-export type KeyToEvent<K extends keyof EventHandlers> = Parameters<Required<EventHandlers>[K]>[0]
+export type KeyToEvent<K extends keyof EventHandlers> = Parameters<
+  Required<EventHandlers>[K]
+>[0]
 
 export type EventHandlers = {
-  onclick?: (event: ThreeEvent<MouseEvent>) => void
-  oncontextmenu?: (event: ThreeEvent<MouseEvent>) => void
-  ondblclick?: (event: ThreeEvent<MouseEvent>) => void
-  onpointerup?: (event: ThreeEvent<PointerEvent>) => void
-  onpointerdown?: (event: ThreeEvent<PointerEvent>) => void
-  onpointerover?: (event: ThreeEvent<PointerEvent>) => void
-  onpointerout?: (event: ThreeEvent<PointerEvent>) => void
-  onpointerenter?: (event: ThreeEvent<PointerEvent>) => void
-  onpointerleave?: (event: ThreeEvent<PointerEvent>) => void
-  onpointermove?: (event: ThreeEvent<PointerEvent>) => void
-  onpointermissed?: (event: MouseEvent) => void
-  onpointercancel?: (event: ThreeEvent<PointerEvent>) => void
-  onwheel?: (event: ThreeEvent<WheelEvent>) => void
+  onclick?: (event: IntersectionEvent<MouseEvent>) => void
+  oncontextmenu?: (event: IntersectionEvent<MouseEvent>) => void
+  ondblclick?: (event: IntersectionEvent<MouseEvent>) => void
+  onpointerup?: (event: IntersectionEvent<PointerEvent>) => void
+  onpointerdown?: (event: IntersectionEvent<PointerEvent>) => void
+  onpointerover?: (event: IntersectionEvent<PointerEvent>) => void
+  onpointerout?: (event: IntersectionEvent<PointerEvent>) => void
+  onpointerenter?: (event: IntersectionEvent<PointerEvent>) => void
+  onpointerleave?: (event: IntersectionEvent<PointerEvent>) => void
+  onpointermove?: (event: IntersectionEvent<PointerEvent>) => void
+  onpointermissed?: (event: IntersectionEvent<PointerEvent>) => void
+  onpointercancel?: (event: IntersectionEvent<PointerEvent>) => void
+  onwheel?: (event: IntersectionEvent<WheelEvent>) => void
 }
