--- src/lib/components/__tests__/AddHandlers.spec.svelte
@@ -1,15 +1,0 @@
-<script>
-  import { Object3D } from 'three'
-  import { useInteractivity } from '@threlte/extras'
-  import AddHandlers from '../AddHandlers.svelte'
-
-  export const interactivity = useInteractivity()
-
-  const ref = new Object3D()
-  ref.name = 'AddHandlers'
-</script>
-
-<AddHandlers
-  {ref}
-  {...$$restProps}
-/>
