--- src/lib/components/__tests__/Container.spec.svelte
@@ -1,16 +1,0 @@
-<svelte:options accessors />
-
-<script>
-  import Container from '../Container/Container.svelte'
-  import Root from '../Root/Root.svelte'
-
-  export let ref
-</script>
-
-<Root>
-  <Container
-    bind:ref
-    name="container"
-    {...$$restProps}
-  />
-</Root>
