+++ src/lib/index.ts
@@ -31,14 +31,14 @@ export type {
 export { type FontFamilies } from '@pmndrs/uikit/internals'
 
 // Components
-export { default as Container } from './components/Container/Container.svelte'
-export { default as Content } from './components/Content/Content.svelte'
-export { default as Fullscreen } from './components/Fullscreen/Fullscreen.svelte'
-export { default as Image } from './components/Image/Image.svelte'
-export { default as Root } from './components/Root/Root.svelte'
-export { default as SVG } from './components/Svg/Svg.svelte'
-export { default as Text } from './components/Text/Text.svelte'
-export { default as Video } from './components/Video/Video.svelte'
+export { default as Container } from './components/Container.svelte'
+export { default as Content } from './components/Content.svelte'
+export { default as Fullscreen } from './components/Fullscreen.svelte'
+export { default as Image } from './components/Image.svelte'
+export { default as Root } from './components/Root.svelte'
+export { default as SVG } from './components/Svg.svelte'
+export { default as Text } from './components/Text.svelte'
+export { default as Video } from './components/Video.svelte'
 
 // Refs
 export type { ComponentInternals } from './useInternals'
